etape 3 fini sans doc ni script instal

+++ sql/produits/1_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Produits (
     idP SERIAL,
     NomP TEXT,
-    Prix INTEGER,
+    Prix NUMERIC,
     type TEXT DEFAULT 'Bague',
     materiaux TEXT DEFAULT 'Or',
     Promo BOOLEAN DEFAULT FALSE,
